% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_scat.R
\name{clip_scat}
\alias{clip_scat}
\title{A function for producing smoothed scatterplots for CLIP-seq data}
\usage{
clip_scat(exp, ctrl, sf_exp = 1, sf_ctrl = 1, xyline = F,
  exp_lab = "crosslinked", ctrl_lab = "non-crosslinked",
  main = "2D density plot", elliptical = 2, nbin = 100,
  colramp = rev(rainbow(10, end = 4/6)))
}
\arguments{
\item{exp}{A vector containing read-counts per nucleotide in a crosslinked sample}

\item{ctrl}{A vector containing read-counts per nucleotide in a non-crosslinked control}

\item{exp_lab}{A label describing the experimental condition}

\item{ctrl_lab}{A label describing the control condition}

\item{elliptical}{The number of standard deviations to use in constructing an elliptical filter}
}
\description{
A function for producing smoothed scatterplots for CLIP-seq data
}
\examples{

}

