% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_scat2.R
\name{clip_scat2}
\alias{clip_scat2}
\title{A function for producing smoothed scatterplots for CLIP-seq data}
\usage{
clip_scat2(exp, ctrl, exp_lab = "crosslinked", ctrl_lab = "non-crosslinked",
  elliptical = 2, nbin = 1000, nrpoints = 0,
  colramp = colorRampPalette(rev(brewer.pal(11, "RdYlBu"))))
}
\arguments{
\item{exp}{A vector containing read-counts per nucleotide in a crosslinked sample}

\item{ctrl}{A vector containing read-counts per nucleotide in a non-crosslinked control}

\item{exp_lab}{A label describing the experimental condition}

\item{ctrl_lab}{A label describing the control condition}

\item{elliptical}{The number of standard deviations to use in constructing an elliptical filter}
}
\description{
A function for producing smoothed scatterplots for CLIP-seq data
}
\examples{

}

