% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_maxima.R
\name{find_maxima}
\alias{find_maxima}
\title{Find local maxima in the empirical distribution of a vector of values}
\usage{
find_maxima(data, n = 2, plot = TRUE)
}
\arguments{
\item{data}{A vector containing numeric data points}

\item{n}{Number of local maxima to find in the empirical distribution}

\item{plot}{Plot density and maxima?}
}
\value{
find_maxima Returns a vector containing n local maxima ordered by
decreasing density.
}
\description{
Find local maxima in the empirical distribution of a vector of values
}
\examples{

}

