% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_minima_in_range.R
\name{find_minima_in_range}
\alias{find_minima_in_range}
\title{A function for finding local minima in the empirical distribution of a vector
of values within a certain range}
\usage{
find_minima_in_range(data, n = 1, range = c(min(data), max(data)),
  plot = TRUE)
}
\arguments{
\item{data}{A vector containing numeric data points}

\item{n}{Number of local minima to find in the empirical distribution}

\item{range}{Range in data within which to find minima}

\item{plot}{Plot density and found minima?}
}
\value{
find_minima Return a vector containing n local minima ordered by
decreasing density.
}
\description{
A function for finding local minima in the empirical distribution of a vector
of values within a certain range
}
\examples{

}

