% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheat.R
\name{ggheat}
\alias{ggheat}
\title{ggplot2 heatmap}
\usage{
ggheat(m, rescaling = "none", clustering = "none", labCol = T,
  labRow = T, border = FALSE, heatscale = c(low = "darkblue", high =
  "lightblue"))
}
\arguments{
\item{m}{A numeric matrix or data.frame}

\item{rescaling}{Rescale? Options are "row" or "column"}

\item{clustering}{Cluster entries? Options are "row", "column", or "both".}

\item{labCol}{Label columns?}

\item{labRow}{Label rows?}

\item{border}{Add border to plot?}

\item{heatscale}{Color range for colorscale.}
}
\value{
A ggplot object containing the formatted heatmap
}
\description{
Draws a (potentially clustered) heatmap, with a lighter-weight resulting
figure file than is typical for heatmap.2. This code adapted from a code
snippet found at:
https://www.r-bloggers.com/ggheat-a-ggplot2-style-heatmap-function/;
if you are the original author please contact me so I can credit you
properly.
}
\details{

}
\examples{

data(mtcars)
x=as.matrix(mtcars)
ggheat(x, clustering='column', rescaling='row', heatscale=c(low='red', high='yellow'))

}
\seealso{
\code{\link{runDiagnostics}}, \code{\link{plotCorrelations}}
}

