% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_scale_factors.R
\name{gm_scale_factors}
\alias{gm_scale_factors}
\title{A function for determining RNA-seq scale factors using the geometric mean
across samples, as introduced by Anders and Huber, Genome Biology, 2010.}
\usage{
gm_scale_factors(matrix)
}
\arguments{
\item{matrix}{A matrix containing columns corresponding to samples, and rows
to per feature read counts}
}
\value{
gm_scale_factor Return an array of scale factors.
}
\description{
A function for determining RNA-seq scale factors using the geometric mean
across samples, as introduced by Anders and Huber, Genome Biology, 2010.
}
\examples{

}

