% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{A function}
\usage{
loadData(data_table)
}
\arguments{
\item{forward_path}{.}

\item{reverse_path}{Path to reverse strand wig file.}
}
\value{
loadData Return a list of IRanges rle coverage vectors.
}
\description{
A function
}
\examples{

}

