% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norclip.R
\name{norclip}
\alias{norclip}
\title{Perform norclip normalization of CLIP-seq libraries}
\usage{
norclip(file, crossnormalize = T, sdn = 8)
}
\arguments{
\item{file}{Path to a description file}
}
\description{
Perform norclip normalization of CLIP-seq libraries
}
\examples{

}

