% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wig2rle.R
\name{wig2rle}
\alias{wig2rle}
\title{A function for importing a wig file and transforming it into a vector
containing per nucleotide counts.}
\usage{
wig2rle(forward_path, reverse_path)
}
\arguments{
\item{forward_path}{Path to forward strand wig file.}

\item{reverse_path}{Path to reverse strand wig file.}
}
\value{
wig2rle Return an IRanges rle coverage vector.
}
\description{
A function for importing a wig file and transforming it into a vector
containing per nucleotide counts.
}
\examples{

}

