% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheat.R
\name{ggheat}
\alias{ggheat}
\title{Calculate scale factors for CLIP-seq data}
\usage{
ggheat(m, rescaling = "none", clustering = "none", labCol = T,
  labRow = T, border = FALSE, heatscale = c(low = "blue", high = "red"))
}
\arguments{
\item{wigs}{A list of IRanges rle coverage vectors.}

\item{data_table}{Vector metadata in data frame format, see package vignette
for details.}

\item{sdn}{Number of standard deviations for elliptical filtering, see
\code{\link{filter_elliptical}} for details.}

\item{crossnormalize}{Crossnormalize the control libraries?}

\item{plot}{Produce diagnostic plots?}

\item{bg_cut}{Cut-off for control library normalization.}
}
\value{
heatmap.2 style plot in ggplots2
}
\description{
Calculate scale factors for CLIP-seq data
}
\details{
stolen from https://www.r-bloggers.com/ggheat-a-ggplot2-style-heatmap-function/
}
\examples{

}
\seealso{
\code{\link{norclip}}, \code{\link{loadData}},
\code{\link{gm_scale_factors}}, \code{\link{filter_elliptical}}
\code{\link{clip_scat}}
}

