% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheat.R
\name{ggheat}
\alias{ggheat}
\title{A heatmap.2 style heatmap in ggplot2}
\usage{
ggheat(m, rescaling = "none", clustering = "none", labCol = T,
  labRow = T, border = FALSE, heatscale = c(low = "blue", high = "red"))
}
\arguments{
\item{m}{A numeric matrix or data.frame}

\item{rescaling}{Rescale? Options are "row" or "column"}

\item{clustering}{Cluster entries? Options are "row", "column", or "both".}

\item{labCol}{Label columns?}

\item{labRow}{Label rows?}

\item{border}{Add border to plot?}

\item{heatscale}{Color range for colorscale.}
}
\value{
heatmap.2 style plot in ggplots2
}
\description{
A heatmap.2 style heatmap in ggplot2
}
\details{
modified from
https://www.r-bloggers.com/ggheat-a-ggplot2-style-heatmap-function/
}
\examples{

}
\seealso{
\code{\link{runDiagnostics}}, \code{\link{plotCorrelations}}
}

